stock added

+++ src/products/products.service.ts
@@ -2,14 +2,15 @@ import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { Product } from './schemas/product.schema';
+import { Stock } from 'src/stock/schemas/stock.schema';
 
 @Injectable()
 export class ProductsService {
   constructor(
-    @InjectModel(Product.name)
-    private productModel: Model<Product>
+    @InjectModel(Product.name) private readonly productModel: Model<Product>,
+    @InjectModel(Stock.name) private readonly stockModel: Model<Stock>,
   ) {}
 
   async create(dto: CreateProductDto) {
@@ -17,8 +18,40 @@ export class ProductsService {
     if (codeExists) {
       throw new BadRequestException('Código ya existe');
     }
-    const createdProduct = await this.productModel.create(dto);
-    return createdProduct.save();
+    // Crear stock;
+    const stock = await this.stockModel.create({ 
+      quantity: dto.initialQuantity ?? 0, 
+      reserved: 0,
+    });
+    console.log('Stock creado:', stock);
+
+    // Crear producto
+    try {
+      const categories: Types.ObjectId[] =
+      (dto.categories ?? []).map(id => new Types.ObjectId(id));
+
+      const product = await this.productModel.create({
+        code: dto.code,
+        name: dto.name,
+        categories: categories,
+        price: dto.price,
+        img_url: dto.img_url,
+        stock: stock._id,
+      });
+
+      console.log('Producto creado:', product);
+
+      return this.productModel
+      .findById(product._id)
+      .populate('stock')
+      .lean()
+      .exec();
+
+    } catch (error) {
+      // Si hay error eliminar el stock
+      await this.stockModel.deleteOne({ _id: stock._id });
+      throw error;
+    }
   }
 
   async createMany(dto: CreateProductDto[]) {
@@ -27,10 +60,6 @@ export class ProductsService {
 
   async findAll(): Promise<Product[]> {
     return this.productModel.find().exec();
-  }
-
-  async findByCategories(filter: any) {
-    return this.productModel.find(filter).lean().exec();
   }
 
   async findOne(id: string) {
@@ -48,6 +77,13 @@ export class ProductsService {
   }
 
   async remove(id: string) {
+    // eliminar el stock asociado 
+    const product = await this.productModel.findById(id).exec();
+    if (!product) {
+      throw new BadRequestException('Producto no encontrado');
+    }
+    const stockId = product.stock;
+    await this.stockModel.findByIdAndDelete({ _id: stockId }).exec();
     return await this.productModel.findByIdAndDelete({ _id: id }).exec();
   }
 }
